Flush cache on page flip for x86
Use clflush_cache_range to flush write-back cache
for page flips, according to Andrew Cooper:
"Unprivileged guests (i.e. ones without hardware)
are not permitted to make mappings with anything
other than a writeback memory type, because
all kinds of chaos can ensue if the guest constructs
aliasing mappings with different cacheabilities,
and it is prohibitively expensive for Xen to track
for auditing purposes."

For that reason we have to have to mmap with PAGE_SHARED
and though have to flush caches.

FIXME: For ARM platform neither applicable that
we can't mmap non-cached pages nor that we have to
provide own mmap.

Signed-off-by: Oleksandr Andrushchenko <Oleksandr_Andrushchenko@epam.com>